tests/wpt/mozilla/tests/mozilla/structuredclone.html

<!doctype html>
<meta charset="utf-8">
<title></title>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<iframe src="resources/iframe_contentDocument_inner.html" id="iframe"></iframe>
<script>
async_test(function(t) {
  var iframe = document.getElementById('iframe');
  var blob = new Blob(['<a id="a"><b id="b">hey!</b></a>'], {type:"text/plain"});
  window.addEventListener("message", this.step_func(function(msg) {
      assert_equals(msg.data.size, blob.size);
      var cloned_content, original_content;
      var reader = new FileReader();
      reader.addEventListener("loadend", this.step_func(function() {
          original_content = reader.result;
          console.log('blob:', reader.result)
          console.log('ok')
      }));
      reader.readAsText(blob);
      var reader2 = new FileReader();
      reader2.addEventListener("loadend", this.step_func_done(function() {
          cloned_content = reader2.result;
          console.log('clone blob:', reader2.result)
          console.log('ok')
          assert_equals(cloned_content, original_content);
      }));
      reader2.readAsText(msg.data);
  }), false);
  iframe.onload = this.step_func(function() {
      window.postMessage(blob, '*');
  }, false);
}, "test cloning blobs");
</script>
</body>
</html>
